Create 24Find The Median.java
Find The Median
The median of a list of numbers is essentially its middle element after sorting. The same number of elements occur after it as before. Given a list of numbers with an odd number of elements, find the median.

Input Format
The first line inputs integer n, size of array a.
The second line inputs n space-seperated integers, i.e., a[i].

Output
In a new line, print the median of the array.

Example 1
Input
7
0 1 2 4 6 5 3

Output
3