Create 25Floyd's Triangle.java
Floyd's Triangle.
Write a program in Java to print the Floyd's Triangle.

Floyd's triangle is a right-angled triangular array of natural numbers, used in computer science education. It is named after Robert Floyd. It is defined by filling the rows of the triangle with consecutive numbers, starting with a 1 in the top left corner:

Input Format
The number of row,

Output Format
Return the rows of the triangle with consecutive numbers, starting with a 1 in the top left corner:

Example 1
Input
5

Output
1 
2 3 
4 5 6 
7 8 9 10 
11 12 13 14 15