Create 26_Water Bill.java
Water Bill
In this task, you need to identify the water bill for a household

You are given the total liter of water used by a house and output the total water bill. The water bill is calculated as follows:

For first 100 litres: Rs. 15 per litre

For next 100 litres: Rs. 14 per litre

After 200 litres: Rs. 12 per litre

Input Format
the first and only line of input is the number of liters of water used

Output Format
return the total bill of the house based on given conditions

Example 1
Input
105

Output
1570